[IMP]Resumen y [FIX] validaciones y migraciones

+++ database/migrations/2021_06_14_224819_create_ventas_table.php
@@ -19,7 +19,7 @@ class CreateVentasTable extends Migration
             $table->integer('piezas');
             $table->date('fecha');
             $table->decimal('precio',$precision = 8, $scale = 2);
-            $table->decimal('total',$precision = 12, $scale = 2);
+            $table->decimal('total',$precision = 16, $scale = 2);
             $table->unsignedBigInteger('cliente_id');
             $table->unsignedBigInteger('estanque_id');
             $table->timestamps();
